+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.6.0...main)
 
+* [#143](https://github.com/mozilla/glean.js/pull/143): Provide a way to initialize and reset Glean.js in tests.
+
 # v0.6.0 (2021-03-22)
 
 * [#123](https://github.com/mozilla/glean.js/pull/123): BUGFIX: Fix support for ES6 environments.